Addition of Phone/Date Mask Showing and Hiding
Added 4 new functions:

events - clears all of the data inputs placeholders if they exist (IU Health seems to have xx/xx/xxxx in place), then sets up all the fields that need to have their masks shown on focus, etc.

clearPlaceholder - sets any input passed to it to have a blank placeholder

showPlaceholder - checks if the events target is 1 of the 2 we are expecting and shows the proper placeholder.

removePlaceholder = removes the placeholder.